+++ activitysampling-application/src/main/java/de/muspellheim/activitysampling/application/ActivityListCell.java
@@ -1,8 +1,16 @@
 package de.muspellheim.activitysampling.application;
 
+import de.muspellheim.activitysampling.domain.*;
+import java.util.function.*;
 import javafx.scene.control.*;
 
 class ActivityListCell extends ListCell<ActivityItem> {
+  private final Consumer<Activity> onSelect;
+
+  ActivityListCell(Consumer<Activity> onSelect) {
+    this.onSelect = onSelect;
+  }
+
   protected void updateItem(ActivityItem item, boolean empty) {
     super.updateItem(item, empty);
 
@@ -10,12 +18,24 @@ class ActivityListCell extends ListCell<ActivityItem> {
       setText(null);
       setGraphic(null);
       setStyle(null);
+      setOnMouseClicked(null);
     } else {
       setText(item.text());
-      if (item.header()) {
-        setStyle("-fx-font-weight: bold");
-      } else {
+      if (item.activity() != null) {
+        // Activity
         setStyle(null);
+        setOnMouseClicked(
+            e -> {
+              if (e.getClickCount() != 2) {
+                return;
+              }
+
+              onSelect.accept(item.activity());
+            });
+      } else {
+        // Group header
+        setStyle("-fx-font-weight: bold");
+        setOnMouseClicked(null);
       }
     }
   }