+++ activitysampling-application/src/test/java/de/muspellheim/activitysampling/application/AcceptanceTests.java
@@ -21,18 +21,53 @@ class AcceptanceTests {
   @Test
   void mainScenario() {
     var eventStore = new CsvEventStore(STORE_FILE);
-    Clock clock = Clock.fixed(Instant.parse("2022-11-16T17:05:00Z"), ZoneId.systemDefault());
+    var clock = new TickingClock(Instant.parse("2022-11-16T17:05:00Z"));
     var activitiesService = new ActivitiesServiceImpl(eventStore, clock);
     var fixture = new ActivitySamplingViewModel(activitiesService);
     fixture.run();
+
+    //
+    // Step 1 - Log first activity
+    //
 
     fixture.activityTextProperty().set("Lorem ipsum");
     fixture.logActivity();
 
     assertEquals(
         List.of(
-            new ActivityItem("Mittwoch, 16. November 2022", true),
-            new ActivityItem("18:05 - Lorem ipsum", false)),
-        fixture.getRecentActivities());
+            new ActivityItem("Mittwoch, 16. November 2022"),
+            new ActivityItem(
+                "18:05 - Lorem ipsum",
+                new Activity(LocalDateTime.of(2022, 11, 16, 18, 5), "Lorem ipsum"))),
+        fixture.getRecentActivities(),
+        "Step 1 - Recent activities");
+
+    //
+    // Step 2 - Log second activity
+    //
+
+    clock.tick(Duration.ofMinutes(20));
+    fixture.activityTextProperty().set("Foobar");
+    fixture.logActivity();
+
+    assertEquals(
+        List.of(
+            new ActivityItem("Mittwoch, 16. November 2022"),
+            new ActivityItem(
+                "18:25 - Foobar", new Activity(LocalDateTime.of(2022, 11, 16, 18, 25), "Foobar")),
+            new ActivityItem(
+                "18:05 - Lorem ipsum",
+                new Activity(LocalDateTime.of(2022, 11, 16, 18, 5), "Lorem ipsum"))),
+        fixture.getRecentActivities(),
+        "Step 2 - Recent activities");
+
+    //
+    // Step 2 - Select first activity
+    //
+
+    clock.tick(Duration.ofMinutes(20));
+    fixture.setActivity(new Activity(LocalDateTime.of(2022, 11, 16, 18, 5), "Lorem ipsum"));
+
+    assertEquals("Lorem ipsum", fixture.activityTextProperty().get(), "Step 3 - Activity text");
   }
 }